Moved out all state logic to a module
