+++ src/state/window.rs
@@ -1,0 +1,8 @@
+use chrono::{DateTime, Utc};
+
+#[derive(Clone, Debug)]
+pub struct Window {
+    pub class: String,
+    pub timestamp: DateTime<Utc>,
+    pub origin: i32,
+}
